Route Rules Migration (Override and Supplier Zone Block)
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C35725

+++ TOneV2/Code/TOneV2/TOne.WhS.DBSync.Entities/Rules/SupplierOption.cs
@@ -9,7 +9,7 @@ namespace TOne.WhS.DBSync.Entities
     {
         public string SupplierId { get; set; }
         public short Priority { get; set; }
-        public short Percentage { get; set; }
+        public decimal Percentage { get; set; }
         public bool IsLoss { get; set; }
 
         public override string ToString()
